P4PU-1257 Moves broker info validation and adds external ID to mapper
