+++ src/main/java/it/gov/pagopa/pu/citizen/connector/organization/client/BrokerSearchClient.java
@@ -2,8 +2,11 @@ package it.gov.pagopa.pu.citizen.connector.organization.client;
 
 import it.gov.pagopa.pu.citizen.connector.organization.config.OrganizationApisHolder;
 import it.gov.pagopa.pu.organization.dto.generated.Broker;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
+@Slf4j
 @Service
 public class BrokerSearchClient {
 
@@ -14,7 +17,14 @@ public class BrokerSearchClient {
   }
 
   public Broker getBrokerByExternalId(String accessToken, String externalId){
-    return organizationApisHolder.getBrokerSearchControllerApi(accessToken)
+
+    try {
+      return organizationApisHolder.getBrokerSearchControllerApi(accessToken)
       .crudBrokersFindBrokerByExternalId(externalId);
+    } catch (
+      HttpClientErrorException.NotFound e) {
+        log.warn("Broker with externalId {} not found",  externalId);
+        return null;
+    }
   }
 }
