+++ src/main/java/it/gov/pagopa/pu/citizen/controller/BrokerController.java
@@ -2,7 +2,6 @@ package it.gov.pagopa.pu.citizen.controller;
 
 import it.gov.pagopa.pu.citizen.controller.generated.BrokerApi;
 import it.gov.pagopa.pu.citizen.dto.generated.BrokerInfoDTO;
-import it.gov.pagopa.pu.citizen.exception.InvalidParamException;
 import it.gov.pagopa.pu.citizen.security.SecurityUtils;
 import it.gov.pagopa.pu.citizen.service.broker.BrokerRetrieverService;
 import lombok.extern.slf4j.Slf4j;
@@ -21,14 +20,7 @@ public class BrokerController implements BrokerApi {
 
   @Override
   public ResponseEntity<BrokerInfoDTO> getBrokerInfo(Long brokerId, String externalId) {
-    log.info("getBrokerInfo was requested with brokerId {} and externalId {}", brokerId, externalId);
-    validateInputFields(brokerId, externalId);
+    log.info("getBrokerInfo was requested with brokerId {} or externalId {}", brokerId, externalId);
     return ResponseEntity.ofNullable(brokerRetrieverService.getBrokerInfo(brokerId, externalId, SecurityUtils.getAccessToken()));
-  }
-
-  private static void validateInputFields(Long brokerId, String externalId) {
-    if (brokerId == null && externalId == null) {
-      throw new InvalidParamException("INVALID_FIELDS","Either brokerId or externalId must be provided");
-    }
   }
 }
