+++ src/main/java/it/gov/pagopa/pu/citizen/service/broker/BrokerRetrieverServiceImpl.java
@@ -3,6 +3,7 @@ package it.gov.pagopa.pu.citizen.service.broker;
 import it.gov.pagopa.pu.citizen.connector.organization.BrokerService;
 import it.gov.pagopa.pu.citizen.connector.organization.OrganizationService;
 import it.gov.pagopa.pu.citizen.dto.generated.BrokerInfoDTO;
+import it.gov.pagopa.pu.citizen.exception.InvalidParamException;
 import it.gov.pagopa.pu.citizen.exception.ResourceNotFoundException;
 import it.gov.pagopa.pu.citizen.mapper.BrokerInfoDTOMapper;
 import it.gov.pagopa.pu.organization.dto.generated.Broker;
@@ -25,6 +26,7 @@ public class BrokerRetrieverServiceImpl implements BrokerRetrieverService {
   @Override
   public BrokerInfoDTO getBrokerInfo(Long brokerId, String externalId, String accessToken) {
     Broker broker;
+    validateInputFields(brokerId, externalId);
     if (brokerId != null){
       broker = brokerService.getBroker(brokerId, accessToken);
     }else {
@@ -34,6 +36,13 @@ public class BrokerRetrieverServiceImpl implements BrokerRetrieverService {
     if(Objects.isNull(broker)){
       throw new ResourceNotFoundException("BROKER_NOT_FOUND","Broker having id "+brokerId+" and " + externalId + "not found");
     }
-    return brokerInfoDTOMapper.map(organizationService.getBrokerOrganization(broker.getBrokerId(),accessToken), broker.getArpuConfig());
+    return brokerInfoDTOMapper.map(organizationService.getBrokerOrganization(broker.getBrokerId(),accessToken), broker.getExternalId(), broker.getArpuConfig());
+  }
+
+
+  private static void validateInputFields(Long brokerId, String externalId) {
+    if (brokerId == null && externalId == null) {
+      throw new InvalidParamException("INVALID_FIELDS","Either brokerId or externalId must be provided");
+    }
   }
 }
